move config and prefetch to sd_task package
